exam_analyzer/requirements.txt
flask
chromadb
sentence-transformers
groq
PyPDF2
python-docx
openpyxl
Pillow
pytesseract
